Move core logic
