+++ crashtracker-ffi/Cargo.toml
@@ -45,10 +45,9 @@ serde = { version = "1.0.214", features = ["derive"] }
 log = "0.4.22"
 
 [target.'cfg(windows)'.dependencies]
-windows = { version = "0.59.0", features = ["Win32_System_Diagnostics_Debug", "Win32_System_Diagnostics_ToolHelp", "Win32_System_ErrorReporting", "Win32_System_Kernel", "Win32_System_ProcessStatus", "Win32_System_Registry", "Win32_System_SystemInformation", "Win32_System_SystemServices", "Win32_System_Threading", "Win32_Security"] }
+windows = { version = "0.59.0", features = ["Win32_System_Diagnostics_Debug", "Win32_System_ErrorReporting"] }
 
 [dev-dependencies]
-goblin = "0.9.3"
 tempfile = "3.3"
 
 [target.'cfg(windows)'.dev-dependencies]
