+++ crashtracker/src/collector_windows/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2021-Present Datadog, Inc. https://www.datadoghq.com/
+// SPDX-License-Identifier: Apache-2.0
+#![cfg(windows)]
+
+pub mod api;
